--- a/pkg/source/uniswapv3/pool_simulator.go
+++ b/pkg/source/uniswapv3/pool_simulator.go
@@ -122,7 +122,7 @@ func NewPoolSimulator(entityPool entity.Pool, chainID valueobject.ChainID) (*Poo
 /**
  * getSqrtPriceLimit get the price limit of pool based on the initialized ticks that this pool has
  */
-func (p *PoolSimulator) getSqrtPriceLimit(zeroForOne bool) *v3Utils.Uint160 {
+func (p *PoolSimulator) getSqrtPriceLimit(zeroForOne bool, result *v3Utils.Uint160) error {
 	var tickLimit int
 	if zeroForOne {
 		tickLimit = p.tickMin
@@ -130,13 +130,13 @@ func (p *PoolSimulator) getSqrtPriceLimit(zeroForOne bool) *v3Utils.Uint160 {
 		tickLimit = p.tickMax
 	}
 
-	sqrtPriceX96Limit, err := v3Utils.GetSqrtRatioAtTickV2(tickLimit)
+	err := v3Utils.GetSqrtRatioAtTickV2(tickLimit, result)
 
 	if err != nil {
-		return nil
+		return err
 	}
 
-	return sqrtPriceX96Limit
+	return nil
 }
 
 func (p *PoolSimulator) CalcAmountIn(param pool.CalcAmountInParams) (*pool.CalcAmountInResult, error) {
@@ -153,8 +153,14 @@ func (p *PoolSimulator) CalcAmountIn(param pool.CalcAmountInParams) (*pool.CalcA
 			tokenOut = p.V3Pool.Token1
 			zeroForOne = true
 		}
+
 		amountOut := coreEntities.FromRawAmount(tokenOut, param.TokenAmountOut.Amount)
-		amountIn, newPoolState, err := p.V3Pool.GetInputAmount(amountOut, p.getSqrtPriceLimit(zeroForOne))
+		var priceLimit v3Utils.Uint160
+		err := p.getSqrtPriceLimit(zeroForOne, &priceLimit)
+		if err != nil {
+			return nil, fmt.Errorf("can not GetInputAmount, err: %+v", err)
+		}
+		amountIn, newPoolState, err := p.V3Pool.GetInputAmount(amountOut, &priceLimit)
 
 		if err != nil {
 			return nil, fmt.Errorf("can not GetInputAmount, err: %+v", err)
@@ -193,45 +199,52 @@ func (p *PoolSimulator) CalcAmountOut(param pool.CalcAmountOutParams) (*pool.Cal
 	tokenOut := param.TokenOut
 	var tokenInIndex = p.GetTokenIndex(tokenAmountIn.Token)
 	var tokenOutIndex = p.GetTokenIndex(tokenOut)
-	var tokenIn *coreEntities.Token
 	var zeroForOne bool
 
 	if tokenInIndex >= 0 && tokenOutIndex >= 0 {
 		if strings.EqualFold(tokenOut, p.V3Pool.Token0.Address.String()) {
 			zeroForOne = false
-			tokenIn = p.V3Pool.Token1
 		} else {
-			tokenIn = p.V3Pool.Token0
 			zeroForOne = true
 		}
-		amountIn := coreEntities.FromRawAmount(tokenIn, tokenAmountIn.Amount)
-		amountOutResult, err := p.V3Pool.GetOutputAmount(amountIn, p.getSqrtPriceLimit(zeroForOne))
+		var amountIn v3Utils.Int256
+		overflow := amountIn.SetFromBig(tokenAmountIn.Amount)
+		if overflow {
+			return nil, ErrOverflow
+		}
+		var priceLimit v3Utils.Uint160
+		err := p.getSqrtPriceLimit(zeroForOne, &priceLimit)
+		if err != nil {
+			return &pool.CalcAmountOutResult{}, fmt.Errorf("can not GetOutputAmount, err: %+v", err)
+		}
+		amountOutResult, err := p.V3Pool.GetOutputAmountV2(&amountIn, zeroForOne, &priceLimit)
 
 		if err != nil {
 			return &pool.CalcAmountOutResult{}, fmt.Errorf("can not GetOutputAmount, err: %+v", err)
 		}
 
 		amountOut := amountOutResult.ReturnedAmount
-		newPoolState := amountOutResult.NewPoolState
+
 		var remainingTokenAmountIn = &pool.TokenAmount{
 			Token: tokenAmountIn.Token,
 		}
 		if amountOutResult.RemainingAmountIn != nil {
-			remainingTokenAmountIn.Amount = amountOutResult.RemainingAmountIn.Quotient()
+			remainingTokenAmountIn.Amount = amountOutResult.RemainingAmountIn.ToBig()
 		} else {
 			remainingTokenAmountIn.Amount = big.NewInt(0)
 		}
+
 		var totalGas = p.gas.BaseGas + p.gas.CrossInitTickGas*int64(amountOutResult.CrossInitTickLoops)
 
 		//p.nextState.SqrtRatioX96 = newPoolState.SqrtRatioX96
 		//p.nextState.Liquidity = newPoolState.Liquidity
 		//p.nextState.TickCurrent = newPoolState.TickCurrent
 
-		if amountOut.Quotient().Cmp(zeroBI) > 0 {
+		if amountOut.Sign() > 0 {
 			return &pool.CalcAmountOutResult{
 				TokenAmountOut: &pool.TokenAmount{
 					Token:  tokenOut,
-					Amount: amountOut.Quotient(),
+					Amount: amountOut.ToBig(),
 				},
 				RemainingTokenAmountIn: remainingTokenAmountIn,
 				Fee: &pool.TokenAmount{
@@ -240,9 +253,9 @@ func (p *PoolSimulator) CalcAmountOut(param pool.CalcAmountOutParams) (*pool.Cal
 				},
 				Gas: totalGas,
 				SwapInfo: UniV3SwapInfo{
-					nextStateSqrtRatioX96: new(uint256.Int).Set(newPoolState.SqrtRatioX96),
-					nextStateLiquidity:    new(uint256.Int).Set(newPoolState.Liquidity),
-					nextStateTickCurrent:  newPoolState.TickCurrent,
+					nextStateSqrtRatioX96: new(uint256.Int).Set(amountOutResult.SqrtRatioX96),
+					nextStateLiquidity:    new(uint256.Int).Set(amountOutResult.Liquidity),
+					nextStateTickCurrent:  amountOutResult.CurrentTick,
 				},
 			}, nil
 		}