--- a/pkg/source/uniswapv3/pool_simulator_test.go
+++ b/pkg/source/uniswapv3/pool_simulator_test.go
@@ -84,6 +84,27 @@ func TestCalcAmountOutConcurrentSafe(t *testing.T) {
 			require.NoError(t, err)
 			_ = result
 		})
+
+		t.Run(tc.name+"new sim", func(t *testing.T) {
+			poolEntity := new(entity.Pool)
+			err := json.Unmarshal([]byte(poolEncoded), poolEntity)
+			require.NoError(t, err)
+
+			poolSim, err := NewPoolSimulator(*poolEntity, valueobject.ChainIDEthereum)
+			require.NoError(t, err)
+
+			result, err := testutil.MustConcurrentSafe[*pool.CalcAmountOutResult](t, func() (any, error) {
+				return poolSim.CalcAmountOut(pool.CalcAmountOutParams{
+					TokenAmountIn: pool.TokenAmount{
+						Token:  tc.tokenIn,
+						Amount: bignumber.NewBig10(tc.amountIn),
+					},
+					TokenOut: tc.tokenOut,
+				})
+			})
+			require.NoError(t, err)
+			_ = result
+		})
 	}
 }
 
@@ -116,6 +137,7 @@ func TestComparePoolSimulatorV2(t *testing.T) {
 			if err == nil {
 				assert.Equal(t, result.TokenAmountOut, resultV2.TokenAmountOut)
 				assert.Equal(t, result.Fee, resultV2.Fee)
+				assert.Equal(t, result.RemainingTokenAmountIn.Amount.String(), resultV2.RemainingTokenAmountIn.Amount.String())
 
 				poolSim.UpdateBalance(pool.UpdateBalanceParams{
 					TokenAmountIn:  in.TokenAmountIn,
@@ -177,6 +199,7 @@ func TestComparePoolSimulatorV2(t *testing.T) {
 			if err == nil {
 				assert.Equal(t, result.TokenAmountOut, resultV2.TokenAmountOut)
 				assert.Equal(t, result.Fee, resultV2.Fee)
+				assert.Equal(t, result.RemainingTokenAmountIn.Amount.String(), resultV2.RemainingTokenAmountIn.Amount.String())
 
 				poolSim.UpdateBalance(pool.UpdateBalanceParams{
 					TokenAmountIn:  in.TokenAmountIn,