--- a/pkg/source/uniswapv3/type.go
+++ b/pkg/source/uniswapv3/type.go
@@ -4,6 +4,10 @@ import (
 	"fmt"
 	"math/big"
 	"strconv"
+
+	"github.com/KyberNetwork/int256"
+	v3Utils "github.com/KyberNetwork/uniswapv3-sdk-uint256/utils"
+	"github.com/holiman/uint256"
 )
 
 type Gas struct {
@@ -12,9 +16,15 @@ type Gas struct {
 }
 
 // UniV3SwapInfo present the after state of a swap
-type UniV3SwapInfo struct {
+type UniV3SwapInfoBigInt struct {
 	nextStateSqrtRatioX96 *big.Int
 	nextStateLiquidity    *big.Int
+	nextStateTickCurrent  int
+}
+
+type UniV3SwapInfo struct {
+	nextStateSqrtRatioX96 *v3Utils.Uint160
+	nextStateLiquidity    *v3Utils.Uint128
 	nextStateTickCurrent  int
 }
 
@@ -59,11 +69,24 @@ type Tick struct {
 	LiquidityNet   *big.Int `json:"liquidityNet"`
 }
 
+type TickU256 struct {
+	Index          int          `json:"index"`
+	LiquidityGross *uint256.Int `json:"liquidityGross"`
+	LiquidityNet   *int256.Int  `json:"liquidityNet"`
+}
+
 type Extra struct {
 	Liquidity    *big.Int `json:"liquidity"`
 	SqrtPriceX96 *big.Int `json:"sqrtPriceX96"`
 	Tick         *big.Int `json:"tick"`
 	Ticks        []Tick   `json:"ticks"`
+}
+
+type ExtraTickU256 struct {
+	Liquidity    *uint256.Int `json:"liquidity"`
+	SqrtPriceX96 *uint256.Int `json:"sqrtPriceX96"`
+	Tick         *int         `json:"tick"`
+	Ticks        []TickU256   `json:"ticks"`
 }
 
 type Slot0 struct {