Benchmark:
**CalcAmountOut**
Before:
```
BenchmarkCalcAmountOut-8                  	   43399	     27103 ns/op	   39050 B/op	     787 allocs/op
BenchmarkCalcAmountOutCrossFewTick-8      	   28562	     41896 ns/op	   57493 B/op	    1172 allocs/op
BenchmarkCalcAmountOutCrossManyTick-8     	    4903	    239083 ns/op	  301602 B/op	    6315 allocs/op
```

After:
```
BenchmarkCalcAmountOutV2-8                	  107085	     11221 ns/op	   12192 B/op	     204 allocs/op
BenchmarkCalcAmountOutCrossFewTickV2-8    	   65976	     18309 ns/op	   18596 B/op	     358 allocs/op
BenchmarkCalcAmountOutCrossManyTickV2-8   	   10000	    105416 ns/op	   97454 B/op	    2309 allocs/op
```

BenchmarkCalcAmountOut: doesn't cross tick
BenchmarkCalcAmountOutCrossFewTick: cross 6 ticks
BenchmarkCalcAmountOutCrossManyTick: cross 79 ticks

**NewPoolSimulator** :
Before:
`BenchmarkNewPoolSimulator-8               	    2325	    516261 ns/op	  242604 B/op	    3778 allocs/op`
After
`BenchmarkNewPoolSimulatorV2-8             	    3356	    363253 ns/op	  147010 B/op	    1606 allocs/op`


## Why did we need it?
<!--- Describe your changes in detail -->

## Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->

## Release Note
<!--- Please list out special notes (if any) when releasing this commit -->
<!--- Special notes is things like breaking changes and related components, required changes and env configurations in kyber-application, etc. -->

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):